LayoutTests/http/tests/quicklook/rtf-document-domain-is-empty-string-expected.txt



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS document.domain is .
